requirements.txt
google-cloud-vision==0.35.2
google-cloud-storage==1.13.2
google-cloud-firestore==0.31.0
Wand==0.5.0
